[spec] Fix create specs for Windows compatibility.
This was a painful exercise.

It appears that using .toMatch() to compare file paths on Windows
causes the Windows's slashes to be interpreted as part of the regex.
This results in identical strings not matching. Our solution is to use
toEqual() but we suspect that we could also escape the backslashes.