Fix for proxy cross domain problem
Headers wasn't forwarded properly as I could see, and for some reason if host is in header it breaks.
Also, body needs to be ether string or array of strings (not sure for array values).

There is maybe better place to do this fix but I am not nodejs dev so it takes me time to connect everything.

My suggestion, do not place host into headers and leave body in format that is in ajax. I am not sure where this modifications happen, maybe is bug in some other lib.