[lib] move getIP to lib/util/getIP.
Only public commands should be placed at the root of lib/.

Moving the file exposed a failing test getIP. Only on the first call
will getIP make a socket connection. All future calls use the cached IP.
By moving the file, the tests are run in a different order and the
IP Address is used instead of the socket request. I fixed this by exposing
the cached IP to allow tests to reset it.