[#114] Add push state support
To accomodate apps that leverage the History API, support for serving the
default page when no matching file is found on the file system.
This commit enables that by adding a new `pushState` middleware, that serves
the `index.html` page when no other middleware reponds to the request.
It is enabled by default, but can be disabled by setting the `pushState`
option to `false`. When disabled, the middleware changes to a simple pass-
through.